requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.33.0
controlnet-aux>=0.0.7
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.66.0
accelerate>=0.24.0
safetensors>=0.4.0
xformers>=0.0.20
omegaconf>=2.3.0
